requirements.txt
opencv-python==4.6.0.66
mediapipe
keras==2.9.0
keras-tuner==1.1.3
scikit-learn==1.1.2
numpy==1.23.3
panda
matplotlib==3.6.0
seaborn==0.12.0
yellowbrick==1.5
Django==4.1.2
djangorestframework==3.14.0
django-cors-headers==3.13.0
django-extensions==3.2.1
protobuf==3.20.*
